Rename interfaces to schema
