+++ src/index.ts
@@ -2,6 +2,6 @@ import { Text } from "havarotjs";
 import { transliterate } from "./transliterate";
 import { sequence } from "./sequence";
 import { remove } from "./remove";
-import { Schema } from "./interfaces";
+import { Schema } from "./schema";
 
 export { remove, Schema, sequence, Text, transliterate };
